♻️ 給油履歴編集ページをモーダルに変更
編集専用ページ（/fuel-logs/[fuelLogId]/edit）を削除し、給油履歴一覧ページ上で
モーダルとして編集・削除できるように変更。削除ボタンは更新ボタン下に配置。
